requirements.txt
# ColdTrack Root Dependencies
# Install with: pip install -r requirements.txt

# AWS SDK
boto3==1.34.20
awscli==1.32.20

# Device Simulator
awsiotsdk==1.21.0

# Lambda Function Dependencies
influxdb-client==1.39.0

# Testing
pytest==7.4.4
pytest-cov==4.1.0
pytest-mock==3.12.0

# Development Tools
python-dotenv==1.0.0
black==24.1.1
flake8==7.0.0
mypy==1.8.0

# Utilities
requests==2.31.0
pyyaml==6.0.1
